requirements.txt
playwright
google-genai
supabase
python-dotenv
playwright-stealth
requests
